+++ src/main/java/com/example/compassobank/entity/ContaPessoal.java
@@ -3,18 +3,19 @@ package com.example.compassobank.entity;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Data
 @Entity
 @Table(name = "ContaPessoal")
 public class ContaPessoal extends Conta{
 
-    private double limite;
+    private BigDecimal limite;
 
     private int cheques;
 
-    private double credito;
+    private BigDecimal credito;
 
-    private double saldoAplicacao;
+    private BigDecimal saldoAplicacao;
 
 }
